requirements.txt
sh
click<=6.6
gdapi-python

